Fixed string parameter conversion for OC
OpenComputers returns strings as byte arrays from `Arguments.checkAny` (because that's what Lua spits out), so they have to be converted to actual strings first to be used as such. Usually this is done via `Arguments.checkString`, but in the current implementation all arguments are automatically parsed into an array using checkAny, so this conversion has to be applied manually (because checkAny intentionally does not change the underlying value).

As it currently is, some methods can not be used properly, because the passed strings will arrive as byte arrays in the underlying implementation, which will then error.

Note that this is *not tested*! It should work, but consider this a suggestion instead of a patch until somebody actually tests it ;-)